Bugfix for #415: rewrite of load_consumption
The day of week weighting guides the contribution of two arrrays.
One is an average of all 7 days.
The other is from what it was last week on that particular day.

Whilst they are the same length (and there is a check for that), they start at different times so all the final summations are off. This fix rectifies that.

refactor: move EV substraction to separate def, no code changes.

Bugfix for #424, add "_" to calls to host  to sunsynk.py to match solis.py factor.